+++ apps/web/lib/utils/locale.test.ts
@@ -1,9 +1,8 @@
 import * as nextHeaders from "next/headers";
 import { describe, expect, test, vi } from "vitest";
-import { TSurvey } from "@formbricks/types/surveys/types";
 import { AVAILABLE_LOCALES, DEFAULT_LOCALE } from "@/lib/constants";
 import { appLanguages } from "@/lib/i18n/utils";
-import { findMatchingLocale, getWebAppLocale } from "./locale";
+import { findMatchingLocale } from "./locale";
 
 // Mock the Next.js headers function
 vi.mock("next/headers", () => ({
@@ -105,87 +104,5 @@ describe("locale", () => {
 
     expect(result).toBe("sv-SE");
     expect(nextHeaders.headers).toHaveBeenCalled();
-  });
-
-  describe("getWebAppLocale", () => {
-    const createMockSurvey = (languages: TSurvey["languages"] = []): TSurvey => {
-      return {
-        id: "survey-1",
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        name: "Test Survey",
-        type: "link",
-        environmentId: "env-1",
-        createdBy: null,
-        status: "draft",
-        displayOption: "displayOnce",
-        autoClose: null,
-        triggers: [],
-        recontactDays: null,
-        displayLimit: null,
-        welcomeCard: {
-          enabled: false,
-          headline: { default: "Welcome" },
-          timeToFinish: false,
-          showResponseCount: false,
-        },
-        questions: [],
-        blocks: [],
-        endings: [],
-        hiddenFields: { enabled: false, fieldIds: [] },
-        variables: [],
-        styling: null,
-        segment: null,
-        languages,
-        displayPercentage: null,
-        isVerifyEmailEnabled: false,
-        isSingleResponsePerEmailEnabled: false,
-        singleUse: null,
-        pin: null,
-        projectOverwrites: null,
-        surveyClosedMessage: null,
-        followUps: [],
-        delay: 0,
-        autoComplete: null,
-        showLanguageSwitch: null,
-        recaptcha: null,
-        isBackButtonHidden: false,
-        isCaptureIpEnabled: false,
-        slug: null,
-        metadata: {},
-      } as TSurvey;
-    };
-
-    test("maps language codes to web app locales", () => {
-      const survey = createMockSurvey();
-      expect(getWebAppLocale("en", survey)).toBe("en-US");
-      expect(getWebAppLocale("de", survey)).toBe("de-DE");
-      expect(getWebAppLocale("pt-BR", survey)).toBe("pt-BR");
-    });
-
-    test("handles 'default' languageCode by finding default language in survey", () => {
-      const survey = createMockSurvey([
-        {
-          language: {
-            id: "lang1",
-            code: "de",
-            alias: null,
-            createdAt: new Date(),
-            updatedAt: new Date(),
-            projectId: "proj1",
-          },
-          default: true,
-          enabled: true,
-        },
-      ]);
-
-      expect(getWebAppLocale("default", survey)).toBe("de-DE");
-    });
-
-    test("falls back to en-US when language is not supported", () => {
-      const survey = createMockSurvey();
-      expect(getWebAppLocale("default", survey)).toBe("en-US");
-      expect(getWebAppLocale("xx", survey)).toBe("en-US");
-    });
   });
 });
